+++ ChangeLog.md
@@ -20,8 +20,10 @@
 - Reworked *Urgent* weapon variant: it now adds priority to the move
 - Reworked *Unfinished Potion* to *Potion Kit*: it now adds a random "potion" consumable to loot (upgrades: Guarding, Urgent)
 - Renamed *Swift* weapon variant to *Accelerating*
-- The Quicken and Slow modifiers now grant/penalize speed scaling with their number of stacks
+- The *Quicken* and *Slow* modifiers now grant/penalize speed scaling with their number of stacks
 - Delvers can now use the default Punch move when they don't have a usable weapon (previously had to have no usable equipment)
+- *Censer*s now gain their bonus when their targets have debuffs (instead of 0 block)
+- Added *Tormenting Censer*
 
 #### Prophets of the Labyrinth Version 0.9.0:
 - Rebalanced predicts, they are now as follows